requirements.txt
# Core dependencies
twisted==23.10.0           # Async networking (optional alternative to socket)
pyyaml==6.0.1             # Config file parsing

# Role B - Intelligence
requests==2.31.0          # HTTP client for APIs
geoip2==4.7.0             # GeoIP lookups (or use ip-api.com)
python-whois==0.8.0       # WHOIS lookups

# Role B - Dashboard
flask==3.0.0              # Web framework
flask-socketio==5.3.5     # Real-time updates
plotly==5.18.0            # Visualizations

# Role B - Alerting
discord-webhook==1.3.0    # Discord integration
slack-sdk==3.23.0         # Slack integration
plyer==2.1.0              # Desktop notifications

# Database
peewee==3.17.0            # ORM for SQLite

# Utilities
python-dotenv==1.0.0      # Environment variables
colorama==0.4.6           # Colored terminal output